# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Приложение для федерации тяжелой атлетики г. Москвы (ФТАМ)
  description: Приложение для спортcменов, тренеров, организаторов и представителей ФТАМ - автоматизация деятельности ФТАМ.
  version: 1.0.11

# servers:
#   - url: https://wlmoscow
...
paths:
...
  /sportsmen/{id}:
    get:
      tags:
        - sportsmen
      summary: Получить спортсмена по ID
      operationId: getSportsmanById
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    patch:
      tags:
        - sportsmen
      summary: Обновление секретарем информации о спортсмене
      description: Обновление информации о сертификате и допуске на соревнования
      operationId: UpdateSportsman
      parameters:
          - name: id
            in: path
            description: ID спортсмена
            required: true
            schema:
              type: string
              format: uuid
      requestBody:
        description: Обновление информации о существующем спортсмене
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSportsman'
        required: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - sportsmen
      summary: Удалить спортсмена
      operationId: deleteSportsman
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
              type: string
              format: uuid
      responses:
        '204':
            description: Successful operation
        '401':
          description: Unauthorized
        '404':
            description: Sportsman not found
        '500':
            description: Internal Server Error
...
  /competitions/{id}/sportsman:
    post:
      tags:
        - competitions
      summary: Запись спортсмена на соревнования
      parameters:
        - name: id
          in: path
          description: ID соревнования
          required: true
          schema:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompApplication'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompApplication'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
  /tcamps/{id}/sportsman:
    post:
      tags:
        - tcamps
      summary: Запись спортсмена на сборы
      parameters:
        - name: id
          in: path
          description: ID сборов
          required: true
          schema:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTCampApplication'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TCampApplication'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    SportsmanExt:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        surname:
          type: string
          example: Иванов
        name:
          type: string
          example: Иван
        patronymic:
          type: string
          example: Иванович
        birthday:
          type: string
          format: date-time
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        gender:
          type: boolean
          example: false
        adoping_validity:
          type: string
          format: date-time
        access_validity:
          type: string
          format: date-time
        access_institution:
          type: string
          example: 1-й диспансер
...
    CompApplication:
      type: object
      properties:
        comp_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        weight_category:
          type: string
          example: 109+
        start_snatch:
          type: integer
          example: 100
        start_clean_and_jerk:
          type: integer
          example: 140
    CreateCompApplication:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        weight_category:
          type: string
          example: 109+
        start_snatch:
          type: integer
          example: 100
        start_clean_and_jerk:
          type: integer
          example: 140
    TCampApplication:
      type: object
      properties:
        tcamp_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    CreateTCampApplication:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
...
    UpdateSportsman:
      type: object
      properties:
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        access_institution:
          type: string
          example: 1-й диспансер
        access_validity:
          type: string
          format: date-time
        antidoping_validity:
          type: string
          format: date-time
...